Continue without error if we're not on a configured channel
The configuration file lists the channels to output to. If we are not on
one of the channels when it comes to outputting an activity, we'd
previously error and not try any further channels.

In addition to potentially missing output to channels, this would cause
us to not reach the point where we record the club_activity_id, leading
to repeatedly trying to output the same activities.

If this happens, just log using irssi_print and try the next channel. Do
not raise an error.