Use single athlete endpoint as club activities endpoint is broken
Apparently the club activities endpoint has been neutered to return
almost none of the data we need (including ID). There is nothing in the
API changelog or docs or the mailing list to indicate whether this is
intentional, but given it's been this way for about a month it looks
like, it seems unlikely to change.

Instead we can get nearly the same data but only for the athlete for
which we have a token. This makes the script work somewhat similarly.

We'll need to do more work to clean up the script (likely ripping parts
of it out, or updating it to support multiple tokens perhaps), but this
gets something working again.